Add files via upload
This notebook shows the following python visualizations: line chart, histogram, bar chart, pie chart and scatter plot. Each visualization is generated by using different libraries - Seaborn, Matplotlib and Pandas. The data is used in the project is "imdb_top_1000.csv" file which is a dataset shared in kaggle showing the top 1000 imdb films. There are a lot of visualization types in Python and this project includes some of them, therefore it is considered as Part 1 in Python Visualizations Repository. The remaining parts will be uploaded as soon as possible.